move store to lib folder
